+++ game_project/entities/characters/player.h
@@ -25,7 +25,7 @@ public:
 	Player(WindowInfo windowInfo, BattlingCharacterType* type, std::string identity,ImageManager& imageManager) : BattlingCharacter(type, identity, imageManager), cardsInHand(1), cardsInDeck(50)
 	{
 		currentTilePos = sf::Vector2i(2, 2);
-		position = sf::Vector2f(currentTilePos.x * windowInfo.tileSizeInPixels, currentTilePos.y * windowInfo.tileSizeInPixels);
+		position = sf::Vector2f(static_cast<float>(currentTilePos.x * windowInfo.tileSizeInPixels), static_cast<float>(currentTilePos.y * windowInfo.tileSizeInPixels));
 		initialise("tic_tac.png", position, imageManager);
 	}
